# Hexagon benchmarks new schedules