Before:
Done, time (conv3x3a16): 0.0049175 s (128 byte mode)
Done, time (dilate3x3): 0.0042597 s (128 byte mode)
Done, time (median3x3): 0.0050552 s (128 byte mode)
Done, time (gaussian5x5): 0.0042595 s (128 byte mode)
Done, time (sobel): 0.0054079 s (128 byte mode)
Done, time (conv3x3a32): 0.0051582 s (128 byte mode)


After:
Done, time (conv3x3a16): 0.0017923 s (128 byte mode)
Done, time (dilate3x3): 0.0019679 s (128 byte mode)
Done, time (median3x3): 0.0021045 s (128 byte mode)
Done, time (gaussian5x5): 0.0021509 s (128 byte mode)
Done, time (sobel): 0.0019982 s (128 byte mode)
Done, time (conv3x3a32): 0.0021427 s (128 byte mode)

@dsharletg, @dpalermo, @ronlieb - There may be better schedules out there that we can explore over time. These are the best I have *so far*.