@pranavb-ca 2017-04-12
I tried by adding `.store_root()` by that in fact degraded performance slightly
Done, time (conv3x3a16): 0.0017827 s (128 byte mode)
Done, time (dilate3x3): 0.0021886 s (128 byte mode)
Done, time (median3x3): 0.0023665 s (128 byte mode)
Done, time (gaussian5x5): 0.0023958 s (128 byte mode)
Done, time (sobel): 0.0021328 s (128 byte mode)
Done, time (conv3x3a32): 0.0021045 s (128 byte mode)

I did not `fold_storage` though. I am trying that right now.