@ronlieb 2017-04-13
The buildbots seemed to mostly pass, one failed building llvm.
There is no way this patch broke the build of llvm.

ok to merge ?